web-server/requirements.txt
Flask
PyJWT==2.3.0
psycopg2-binary
Werkzeug
requests

Pillow==9.5.0
opencv-python
opencv-python-headless
psycopg2-binary

google-cloud-pubsub
google-cloud-logging
google-cloud-storage
